# backend/sonora-openapi.yml
openapi: 3.0.3
info:
  title: Sonora API
  version: 1.0.0
  description: Backend API for Sonora music streaming.
paths:
...
  /api/v1/media/downloads/{token}/:
    get:
      operationId: v1_media_downloads_retrieve
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /api/v1/media/streams/{token}/:
    get:
      operationId: v1_media_streams_retrieve
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT